backend/requirements.txt
asgiref==3.5.2
asttokens==2.2.1
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
bidict==0.22.1
black==22.10.0
Brotli==1.0.9
click==8.1.3
decorator==5.1.1
dj-database-url==1.0.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
executing==1.2.0
gevent==22.10.2
gevent-socketio==0.2.1
gevent-websocket==0.10.1
greenlet==2.0.1
gunicorn==20.1.0
inflection==0.5.1
ipdb==0.13.11
ipython==8.8.0
jedi==0.18.2
jsonschema==4.17.3
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==0.21.0
python-engineio==4.3.4
pytz==2022.6
PyYAML==6.0
six==1.16.0
sphinx-me==0.3
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.8.1
typing==3.7.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
wcwidth==0.2.6
whitenoise==6.2.0
zope.event==4.6
zope.interface==5.5.2
